Tide up asset editor
1st try to center object in AssetEditor. BoundingSphere is not estimated correctly for geometries out of center. Correct the boundingSphere estimation next time.

Bug Preview mode when not logged in. There is a bug in 3D viewing.

Bug for audio in preview mode for Desktop.

Bugs:
Bug: Check robustness when asset names have a space

Other bugs
Audio bug: Silence sound should be removed somehow in Asset Editor. The silence was inserted because the Android wrapper crashes when it does not find audio file to start automatically (this is bug of android actually).

Asset Editor
Uploading enhancement: Add progressbar for uploading

Scene Editor
Improvement: Add progress bar for each item independently